+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.0 - 2020-02-14
+### Changed
+- Reviews are now attributed to purchasables rather than products
+- Review records have a purchasable Id and purchasable Type
+
 ## 0.0.1 - 2019-03-20
 ### Added
 - Initial release